+++ composer.json
@@ -7,7 +7,7 @@
     },
     "autoload": {
         "psr-4": {
-            "App\\": "src/"
+            "App\\": ""
         }
     }
 }
