+++ config/database.php
@@ -1,0 +1,9 @@
+<?php
+
+return [
+    'mongo' => [
+        'domain' => 'localhost',
+        'port' => 27017,
+        'database' => 'notes'
+    ]
+];
