+++ config/mongo.php
@@ -1,0 +1,7 @@
+<?php
+
+return [
+    'domain' => 'localhost',
+    'port' => 27017,
+    'database' => 'notes'
+];
