+++ config/user.php
@@ -1,0 +1,10 @@
+<?php
+
+return [
+    'auth' => [
+        'password' => [
+            'length' => 8
+        ],
+        'salt' => 'ghb5e7rdyufnq3g47b'
+    ]
+];
