+++ public/.htaccess
