--- src/config/_autoload.php
@@ -1,8 +1,0 @@
-<?php
-
-/**
- *	'system.php' unavailable 
- */
-require_once 'auth.php';
-require_once 'mongo.php';
-require_once 'app.php';
