--- src/config/app.php
@@ -1,3 +1,0 @@
-<?php
-
-$config['dir']['root'] = __DIR__ . '/../../';
