--- src/config/mongo.php.sample
@@ -1,5 +1,0 @@
-<?php
-
-$config['mongodb']['domain'] = '';
-$config['mongodb']['port'] = 0;
-$config['mongodb']['dbname'] = '';
