--- src/messages/_autoload.php
@@ -1,7 +1,0 @@
-<?php
-
-require 'auth.php';
-require 'api.php';
-require 'app.php';
-require 'user.php';
-require 'mongo.php';
