--- src/messages/api.php
@@ -1,6 +1,0 @@
-<?php
-
-$messages['api']['user']['create']['error'] = 'User creating error';
-$messages['api']['user']['get']['error'] = 'User getting error';
-$messages['api']['user']['validate']['error'] = 'User validating error';
-$messages['api']['version']['error'] = 'API version was not set';
