--- src/messages/app.php
@@ -1,80 +1,0 @@
-<?php
-
-/* just for fun */
-$messages['welcome'] = "<pre style='font:10px/5px monospace;'>
-                                  ```.```                                  
-                          `+@@@@@@@@@@@@@@@@@@@+.                          
-                      .#@@@@+:               ,+@@@@@,                      
-                   ,@@@@,     `:+@@@@@@@@@+;`     .#@@@;                   
-                 @@@#    :@@@@@@@@@@@@@@@@@@@@@@@;    '@@@`                
-              ,@@@   `#@@@@@@@@@+;,..``.,;+@@@@@@@@@@.   #@@:              
-            `@@+   #@@@@@@+`                    '@@@@@@@   '@@,            
-           @@'  ,@@@@@@`                            #@@@@@:  ;@@`          
-         +@#  :@@@@@`                                  @@@@@'  +@#         
-        @@` .@@@@+                                       ;@@@@:  @@`       
-      .@#  @@@@;  .                     @@@+,           `  ,@@@@  +@;      
-      +@ .@@@+ `'            :#@@@@@'.  @@@@@@@@;         :, ;@@@; +@      
-      +@ @@@ `@          ,@@@@@@@@@@@@@@#@@@@@@@@@@.        #. @@@ +@      
-      +@ @@ @,         #@@@@@#;;;;;'#@@@@@@;;;@@@@@@@,       `@ @@ +@      
-      ;@ @@ @         .@@@';;,..........;@@#..:;;@@@@@@       # @@ #@      
-      :@ @@.'        .;#@@@@'..............:#.....:@@@@@#     . @@ @#      
-      .@ @@:     '@@@@@@@@':.........................@@@@@.     @@ @;      
-       @ @,   +@@@@@@;................................,@@@@@   `@@ @,      
-      `@    @@@@@#,.........,,;+@@@',...................'@@@@`     @@;     
-     '@'  #@@@@;........;;;@@@'........,'+';..............@@@@@     .@     
-    #@` `@@@@'...,:;;;;;;;;;'@,.....'++@@@@@@@@'...........'@@@@@@   @     
-   +@  '@@@@,'#@@@@@@@@#:;;;;;;:,........;#@@@@@@@......,....@@@@@  `@@    
-  :@  +@@@@@@@@@@@@;...................;;#@@@@+:;@@@,.:...;;;;@@@,    @@,  
-  @, :@@@@@@@@@@@@@@';;;;:,.............;;+@@@#::':#@@:.'@@@@@@@@@@@:  :@` 
- @#     `@@@@@@@@@@@@@@@@';:.............;;;@@@@+'@@@@@@@@@@@@@@@@@@@@  '@ 
-.@@@@@@+  #@@@@@@@@@@@@@';;;;;,............;;;#@@@@@@@@@@@@@@@@'::+@@@@  @,
-    ,@; `@@@@@@@@@@@@@@@@@@@;;;;;;.:;,........,,:;#@@@@@@@@@@@':::::@@@` #@
-   `@, '@@@@@@@@@@@@@@@@@@@@@@@;;;;;:@@;;,.,......;'';;@@@#::::::::::@@; ;@
-   @+ '@@@@@@@@@@@@@@@@@@@@@@@@@@@;;;+@@@;;;;:;;;;;;'@@@#::::::::::::@@; '@
-  #@ .@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@;;@@@@#;;';;+@@@@@@::::::;':::::;@@` @@
-  @  @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@;;@'@@@@@@::;@@@@@@@@:::@@@  @:
- +@ +@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@'@@@@@@+:@@@@@@@@@@@@:'@@+  @ 
- @: @@:      #@@@@@@@@@@@@@@@@+:@@@@@@@@@@@@@@@@@@@+@@@@@#`   `@@@+@@@  @@ 
- @     .+@, #@@@@@@@@@@@@@@@@@@;.:@@@@@@@@@@@@@@@@@@@@@,       :@@@@@  +@  
-`@  '@@@@@ :@@@@@@@@@@@@@@@;..:;':,@@#@@@@@@@@@@@@@@@.          @@@.  #@   
-,@@@@;  @' @@@@@@@@@@@@@@@@@#;:.....'@@;;+@@@@#+@@@@@@@@@:      #    @@    
-:@#     @ ,@@@@@@@@@@@@@@@@@@@@;;;;:,.,+@+;,..,@@@@@@@@@@@@       `@@@     
-       .@ @@@@+,  :@@@@@@@@@@@@@@@@:........;,@@@@@@@@@@@@'   ,#@@@#       
-       '@ @,      @@@@@@@@@@@@@@;:::;;;',:;;'@@@@@@@@@@,     :@@'          
-       #@        @@@@@@@@@@@@@@@@@@@@@@;;;;@@@@@@@@@@     @@@@@;           
-       #@        @@,        :@@@@@@@@@@@@@@@@@@@@@@`    @@@     '@@@@@@    
-       ;@@                '@@@@@@@@@@@@@@@@@@@@@@+      #@@   ;@@+++++@    
-        @@               @@@@@@@@@@@@@@@@@@@@@@@          #@ `@:      @    
-        @  @@@@@@@@@;           .#@@@@@@@@@@@@      :@@@@` @+@' +@@@. @    
-       @@ @@@@@@@@@@;                :@@@@@         :@@@@@ `@@ '@@@@. @    
-       @ `@@@@@@@@@@;                  @@@@         :@@@@@@ #  @@@@@. @    
-      #@ @@@@                          @@@@            :@@@,  @@@@    @    
-      @  @@@.                      ````@@@@             @@@@ #@@@: ,..@    
-     #@ @@@@          @@@@@@@@   ;@@@@@@@@@  @@@@@@@:    @@@@@@@@ #@@@@    
-     @` @@@,   ```   @@@@@@@@@@  @@@@@@@@@@ @@@@@@@@: `  :@@@@@@ `@        
-    '@ @@@@   `@@@   @@@@@@@@@@  @@@###@@@@ @@@@@@@@: #@  @@@@@. @+        
-    @, @@@+   @@@@   @@@:  `@@@  @@@   @@@@ @@@#      #@  ;@@@@ #@         
-    @# @@@@    @@@   @@@:  `@@@  @@@   @@@@ @@@@@@@   ##  @@@@@` @:        
-     @  @@@@         @@@:  `@@@  @@@   @@@@ @@@@@@@      @@@@@@@ :@        
-     #@ #@@@`        @@@,  `@@@  @@@   @@@@ @@@@@@@     @@@@;@@@+ @@       
-      @. @@@@        @@@,  `@@@  @@@   @@@@ @@@#       :@@@@ '@@@  @@@@    
-      +@ '@@@:`,'+@' @@@@@@@@@@  @@@@@@@@@@ @@@@@@@@: @@@@@   @@@@    @    
-       @; @@@@@@@@@' @@@@@@@@@@  @@@@@@@@@. ;@@@@@@@: @@@@  @  @@@@+` @    
-       ,@ .@@@@@@@@'  @@@@@@@#    @@@@@@@    ,@@@@@@: @@@' @@@ #@@@@. @    
-        @# ;@@@@#',`                                  `,  '@ @, @@@@. @    
-         @;          .:'#                         +':.  .@@` ;@  :#@. @    
-         :@@#@@@@@@@@@@@@@ :@@,              @@# +@@@@@@@#    #@'     @    
-           ,@+;.         @# +@@` '       ;  @@@ :@             `@@@@@#@    
-                          @; @@@  +     #  @@@ .@.                 .;@@    
-                          .@. @@@  :' :+  @@@  @;                          
-                           ;@` @@@` .@:  @@@  @+                           
-                            ;@` @@@`    @@@  @#                            
-                             +@` @@@: `@@@  @#                             
-                              '@. #@@+@@@  @#                              
-                               ;@, '@@@# `@#                               
-                                ,@' ,@' ,@+                                
-                                 `@#   +@,                                 
-                                   @@ @@`                                  
-                                    @@@                                    
-                                     +                                     
-</pre>";
