--- src/messages/auth.php
@@ -1,6 +1,0 @@
-<?php
-
-$messages['auth']['password']['empty'] = 'Password is empty';
-$messages['auth']['password']['length'] = 'Password length is not %d';
-$messages['auth']['userId']['length'] = 'User id length is not %d';
-$messages['auth']['userId']['empty'] = 'User ID is empty';
