--- src/messages/mongo.php
@@ -1,4 +1,0 @@
-<?php
-
-$messages['mongo']['init']['error'] = 'Mongo int connection faulted';
-$messages['mongo']['collection']['get']['error'] = 'Collection %s is not received';
