--- src/messages/user.php
@@ -1,3 +1,0 @@
-<?php
-
-$messages['user']['init']['fault'] = 'User fault init';
