--- src/modules/Http.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace App\Modules;
-
-class HTTP
-{
-    CONST CODE_SERVER_ERROR = 500;
-    CONST CODE_BAD_REQUEST = 400;
-    CONST CODE_SUCCESS = 200;
-}
