+++ system/Http.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\System;
+
+class HTTP
+{
+    CONST CODE_SERVER_ERROR = 500;
+    CONST CODE_BAD_REQUEST = 400;
+    CONST CODE_SUCCESS = 200;
+}
