+++ system/utilities/Base.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\System\Utilities;
+
+class Base {
+    CONST DIR_CONFIG   = 'config';
+    CONST DIR_SYSTEM   = 'system';
+    CONST DIR_VENDOR   = 'vendor';
+    CONST DIR_LOGS     = 'logs';
+    CONST DIR_VERSIONS = 'versions';
+    CONST DIR_MESSAGES = 'messages';
+}
