+++ versions/v1/controllers/Base.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Versions\V1\Controllers;
+
+class Base {
+
+    protected $apiVer = 'v1';
+
+    public function __construct()
+    {
+
+    }
+}
