+++ versions/v1/messages/app.php
@@ -1,0 +1,79 @@
+<?php
+
+return ['welcome' => "<pre style='font:10px/5px monospace;'>
+                                  ```.```                                  
+                          `+@@@@@@@@@@@@@@@@@@@+.                          
+                      .#@@@@+:               ,+@@@@@,                      
+                   ,@@@@,     `:+@@@@@@@@@+;`     .#@@@;                   
+                 @@@#    :@@@@@@@@@@@@@@@@@@@@@@@;    '@@@`                
+              ,@@@   `#@@@@@@@@@+;,..``.,;+@@@@@@@@@@.   #@@:              
+            `@@+   #@@@@@@+`                    '@@@@@@@   '@@,            
+           @@'  ,@@@@@@`                            #@@@@@:  ;@@`          
+         +@#  :@@@@@`                                  @@@@@'  +@#         
+        @@` .@@@@+                                       ;@@@@:  @@`       
+      .@#  @@@@;  .                     @@@+,           `  ,@@@@  +@;      
+      +@ .@@@+ `'            :#@@@@@'.  @@@@@@@@;         :, ;@@@; +@      
+      +@ @@@ `@          ,@@@@@@@@@@@@@@#@@@@@@@@@@.        #. @@@ +@      
+      +@ @@ @,         #@@@@@#;;;;;'#@@@@@@;;;@@@@@@@,       `@ @@ +@      
+      ;@ @@ @         .@@@';;,..........;@@#..:;;@@@@@@       # @@ #@      
+      :@ @@.'        .;#@@@@'..............:#.....:@@@@@#     . @@ @#      
+      .@ @@:     '@@@@@@@@':.........................@@@@@.     @@ @;      
+       @ @,   +@@@@@@;................................,@@@@@   `@@ @,      
+      `@    @@@@@#,.........,,;+@@@',...................'@@@@`     @@;     
+     '@'  #@@@@;........;;;@@@'........,'+';..............@@@@@     .@     
+    #@` `@@@@'...,:;;;;;;;;;'@,.....'++@@@@@@@@'...........'@@@@@@   @     
+   +@  '@@@@,'#@@@@@@@@#:;;;;;;:,........;#@@@@@@@......,....@@@@@  `@@    
+  :@  +@@@@@@@@@@@@;...................;;#@@@@+:;@@@,.:...;;;;@@@,    @@,  
+  @, :@@@@@@@@@@@@@@';;;;:,.............;;+@@@#::':#@@:.'@@@@@@@@@@@:  :@` 
+ @#     `@@@@@@@@@@@@@@@@';:.............;;;@@@@+'@@@@@@@@@@@@@@@@@@@@  '@ 
+.@@@@@@+  #@@@@@@@@@@@@@';;;;;,............;;;#@@@@@@@@@@@@@@@@'::+@@@@  @,
+    ,@; `@@@@@@@@@@@@@@@@@@@;;;;;;.:;,........,,:;#@@@@@@@@@@@':::::@@@` #@
+   `@, '@@@@@@@@@@@@@@@@@@@@@@@;;;;;:@@;;,.,......;'';;@@@#::::::::::@@; ;@
+   @+ '@@@@@@@@@@@@@@@@@@@@@@@@@@@;;;+@@@;;;;:;;;;;;'@@@#::::::::::::@@; '@
+  #@ .@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@;;@@@@#;;';;+@@@@@@::::::;':::::;@@` @@
+  @  @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@;;@'@@@@@@::;@@@@@@@@:::@@@  @:
+ +@ +@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@'@@@@@@+:@@@@@@@@@@@@:'@@+  @ 
+ @: @@:      #@@@@@@@@@@@@@@@@+:@@@@@@@@@@@@@@@@@@@+@@@@@#`   `@@@+@@@  @@ 
+ @     .+@, #@@@@@@@@@@@@@@@@@@;.:@@@@@@@@@@@@@@@@@@@@@,       :@@@@@  +@  
+`@  '@@@@@ :@@@@@@@@@@@@@@@;..:;':,@@#@@@@@@@@@@@@@@@.          @@@.  #@   
+,@@@@;  @' @@@@@@@@@@@@@@@@@#;:.....'@@;;+@@@@#+@@@@@@@@@:      #    @@    
+:@#     @ ,@@@@@@@@@@@@@@@@@@@@;;;;:,.,+@+;,..,@@@@@@@@@@@@       `@@@     
+       .@ @@@@+,  :@@@@@@@@@@@@@@@@:........;,@@@@@@@@@@@@'   ,#@@@#       
+       '@ @,      @@@@@@@@@@@@@@;:::;;;',:;;'@@@@@@@@@@,     :@@'          
+       #@        @@@@@@@@@@@@@@@@@@@@@@;;;;@@@@@@@@@@     @@@@@;           
+       #@        @@,        :@@@@@@@@@@@@@@@@@@@@@@`    @@@     '@@@@@@    
+       ;@@                '@@@@@@@@@@@@@@@@@@@@@@+      #@@   ;@@+++++@    
+        @@               @@@@@@@@@@@@@@@@@@@@@@@          #@ `@:      @    
+        @  @@@@@@@@@;           .#@@@@@@@@@@@@      :@@@@` @+@' +@@@. @    
+       @@ @@@@@@@@@@;                :@@@@@         :@@@@@ `@@ '@@@@. @    
+       @ `@@@@@@@@@@;                  @@@@         :@@@@@@ #  @@@@@. @    
+      #@ @@@@                          @@@@            :@@@,  @@@@    @    
+      @  @@@.                      ````@@@@             @@@@ #@@@: ,..@    
+     #@ @@@@          @@@@@@@@   ;@@@@@@@@@  @@@@@@@:    @@@@@@@@ #@@@@    
+     @` @@@,   ```   @@@@@@@@@@  @@@@@@@@@@ @@@@@@@@: `  :@@@@@@ `@        
+    '@ @@@@   `@@@   @@@@@@@@@@  @@@###@@@@ @@@@@@@@: #@  @@@@@. @+        
+    @, @@@+   @@@@   @@@:  `@@@  @@@   @@@@ @@@#      #@  ;@@@@ #@         
+    @# @@@@    @@@   @@@:  `@@@  @@@   @@@@ @@@@@@@   ##  @@@@@` @:        
+     @  @@@@         @@@:  `@@@  @@@   @@@@ @@@@@@@      @@@@@@@ :@        
+     #@ #@@@`        @@@,  `@@@  @@@   @@@@ @@@@@@@     @@@@;@@@+ @@       
+      @. @@@@        @@@,  `@@@  @@@   @@@@ @@@#       :@@@@ '@@@  @@@@    
+      +@ '@@@:`,'+@' @@@@@@@@@@  @@@@@@@@@@ @@@@@@@@: @@@@@   @@@@    @    
+       @; @@@@@@@@@' @@@@@@@@@@  @@@@@@@@@. ;@@@@@@@: @@@@  @  @@@@+` @    
+       ,@ .@@@@@@@@'  @@@@@@@#    @@@@@@@    ,@@@@@@: @@@' @@@ #@@@@. @    
+        @# ;@@@@#',`                                  `,  '@ @, @@@@. @    
+         @;          .:'#                         +':.  .@@` ;@  :#@. @    
+         :@@#@@@@@@@@@@@@@ :@@,              @@# +@@@@@@@#    #@'     @    
+           ,@+;.         @# +@@` '       ;  @@@ :@             `@@@@@#@    
+                          @; @@@  +     #  @@@ .@.                 .;@@    
+                          .@. @@@  :' :+  @@@  @;                          
+                           ;@` @@@` .@:  @@@  @+                           
+                            ;@` @@@`    @@@  @#                            
+                             +@` @@@: `@@@  @#                             
+                              '@. #@@+@@@  @#                              
+                               ;@, '@@@# `@#                               
+                                ,@' ,@' ,@+                                
+                                 `@#   +@,                                 
+                                   @@ @@`                                  
+                                    @@@                                    
+                                     +                                     
+</pre>"];
