+++ versions/v1/messages/mongo.php
@@ -1,0 +1,12 @@
+<?php
+
+return [
+    'init' => [
+        'error' => 'Mongo int connection faulted'
+    ],
+    'collection' => [
+        'get' => [
+            'error' => 'Collection %s is not received'
+        ]
+    ]
+];
