+++ versions/v1/messages/user.php
@@ -1,0 +1,20 @@
+<?php
+
+return [
+    'isset' => [
+        'notIsset' => 'User not isset'
+    ],
+    'id' => [
+        'empty' => 'User id is empty',
+        'length' => 'User id length is not %d'
+    ],
+    'password' => [
+        'empty' => 'Password is empty'
+    ],
+    'create' => [
+        'error' => 'User creating error'
+    ],
+    'validate' => [
+        'not' => 'User credentials not validated'
+    ]
+];
